refactor: merge 'useEstimations' into 'useStaking'
